+++ NEWS.md
@@ -1,3 +1,5 @@
+* `gather()` throws an error if a column is a data frame (#553)
+
 # tidyr 0.8.3
 
 * `crossing()` preserves factor levels (#410), now works with list-columns