+++ NEWS.md
@@ -1,3 +1,5 @@
+# tidyr (development version)
+
 * `gather()` throws an error if a column is a data frame (#553)
 
 # tidyr 0.8.3